app/requirements.txt
asgiref==3.6.0
cachetools==4.2.4
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
cssbeautifier==1.14.7
Django==4.2.1
django-admin-interface==0.26.0
django-colorfield==0.8.0
django-googledrive-storage==1.6.0
django-widget-tweaks==1.4.12
djlint==1.29.0
EditorConfig==0.12.3
google-api-core==2.10.2
google-api-python-client==2.86.0
google-auth==1.35.0
google-auth-httplib2==0.1.0
googleapis-common-protos==1.59.0
gunicorn==20.1.0
html-tag-names==0.1.2
html-void-elements==0.1.0
httplib2==0.22.0
idna==3.4
jsbeautifier==1.14.7
json5==0.9.14
pathspec==0.11.1
Pillow==9.5.0
protobuf==4.23.0
psycopg2==2.9.6
psycopg2-binary==2.9.6
pyasn1==0.5.0
pyasn1-modules==0.3.0
pyparsing==3.0.9
python-dateutil==2.8.2
python-decouple==3.8
python-slugify==8.0.1
PyYAML==6.0
regex==2023.5.5
requests==2.30.0
rsa==4.9
six==1.16.0
sqlparse==0.4.4
text-unidecode==1.3
tomli==2.0.1
tqdm==4.65.0
uritemplate==4.1.1
urllib3==2.0.2
whitenoise==6.4.0
